requirements.txt
transformers>=4.43.0
accelerate>=0.33.0
bitsandbytes>=0.43.1
datasets>=2.20.0
sentencepiece>=0.1.99
pandas>=2.2.0
numpy>=1.26.0
plotly>=5.22.0
sacrebleu>=2.4.2
tiktoken>=0.7.0
psutil>=5.9.0
PyYAML>=6.0.1
openai>=1.40.0
